+++ lib/docusign_esign/api/accounts_api.rb
@@ -141,6 +141,12 @@ module DocuSign_eSign
     attr_accessor :start_position
 
     # 
+    attr_accessor :task_source
+
+    # 
+    attr_accessor :task_type
+
+    # 
     attr_accessor :user_name_substring
 
     def self.default
@@ -217,6 +223,12 @@ module DocuSign_eSign
 
     # 
     attr_accessor :start_position
+
+    # 
+    attr_accessor :task_source
+
+    # 
+    attr_accessor :task_type
 
     # 
     attr_accessor :user_name_substring
@@ -1727,6 +1739,8 @@ module DocuSign_eSign
       query_params[:'include_closed_users'] = options.include_closed_users if !options.include_closed_users.nil?
       query_params[:'permissions'] = options.permissions if !options.permissions.nil?
       query_params[:'start_position'] = options.start_position if !options.start_position.nil?
+      query_params[:'task_source'] = options.task_source if !options.task_source.nil?
+      query_params[:'task_type'] = options.task_type if !options.task_type.nil?
       query_params[:'user_name_substring'] = options.user_name_substring if !options.user_name_substring.nil?
 
       # header parameters
@@ -2129,108 +2143,6 @@ module DocuSign_eSign
         :auth_names => auth_names)
       if @api_client.config.debugging
         @api_client.config.logger.debug "API called: AccountsApi#get_brand_resources_by_content_type\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
-      end
-      return data, status_code, headers
-    end
-
-    # Gets completion rate for the template
-    # 
-    # @param account_id The external account number (int) or account ID Guid.
-    # @param template_id The ID of the template being accessed.
-    # @return [TemplateCompletionRateResponse]
-    def get_completion_rate_for_template(account_id, template_id)
-      data, _status_code, _headers = get_completion_rate_for_template_with_http_info(account_id, template_id)
-      return data
-    end
-
-    # Gets completion rate for the template
-    # 
-    # @param account_id The external account number (int) or account ID Guid.
-    # @param template_id The ID of the template being accessed.
-    # @return [Array<(TemplateCompletionRateResponse, Fixnum, Hash)>] TemplateCompletionRateResponse data, response status code and response headers
-    def get_completion_rate_for_template_with_http_info(account_id, template_id)
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "Calling API: AccountsApi.get_completion_rate_for_template ..."
-      end
-      # verify the required parameter 'account_id' is set
-      fail ArgumentError, "Missing the required parameter 'account_id' when calling AccountsApi.get_completion_rate_for_template" if account_id.nil?
-      # verify the required parameter 'template_id' is set
-      fail ArgumentError, "Missing the required parameter 'template_id' when calling AccountsApi.get_completion_rate_for_template" if template_id.nil?
-      # resource path
-      local_var_path = "/v2.1/accounts/{accountId}/templates/{templateId}/insights/completionRate".sub('{format}','json').sub('{' + 'accountId' + '}', account_id.to_s).sub('{' + 'templateId' + '}', template_id.to_s)
-
-      # query parameters
-      query_params = {}
-
-      # header parameters
-      header_params = {}
-      # HTTP header 'Accept' (if needed)
-      header_params['Accept'] = @api_client.select_header_accept(['application/json'])
-
-      # form parameters
-      form_params = {}
-
-      # http body (model)
-      post_body = nil
-      auth_names = []
-      data, status_code, headers = @api_client.call_api(:GET, local_var_path,
-        :header_params => header_params,
-        :query_params => query_params,
-        :form_params => form_params,
-        :body => post_body,
-        :auth_names => auth_names,
-        :return_type => 'TemplateCompletionRateResponse')
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "API called: AccountsApi#get_completion_rate_for_template\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
-      end
-      return data, status_code, headers
-    end
-
-    # Gets completion rate for the user
-    # 
-    # @param account_id The external account number (int) or account ID Guid.
-    # @return [UserCompletionRateResponse]
-    def get_completion_rate_for_template_owner(account_id)
-      data, _status_code, _headers = get_completion_rate_for_template_owner_with_http_info(account_id)
-      return data
-    end
-
-    # Gets completion rate for the user
-    # 
-    # @param account_id The external account number (int) or account ID Guid.
-    # @return [Array<(UserCompletionRateResponse, Fixnum, Hash)>] UserCompletionRateResponse data, response status code and response headers
-    def get_completion_rate_for_template_owner_with_http_info(account_id)
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "Calling API: AccountsApi.get_completion_rate_for_template_owner ..."
-      end
-      # verify the required parameter 'account_id' is set
-      fail ArgumentError, "Missing the required parameter 'account_id' when calling AccountsApi.get_completion_rate_for_template_owner" if account_id.nil?
-      # resource path
-      local_var_path = "/v2.1/accounts/{accountId}/templates/insights/completionRate".sub('{format}','json').sub('{' + 'accountId' + '}', account_id.to_s)
-
-      # query parameters
-      query_params = {}
-
-      # header parameters
-      header_params = {}
-      # HTTP header 'Accept' (if needed)
-      header_params['Accept'] = @api_client.select_header_accept(['application/json'])
-
-      # form parameters
-      form_params = {}
-
-      # http body (model)
-      post_body = nil
-      auth_names = []
-      data, status_code, headers = @api_client.call_api(:GET, local_var_path,
-        :header_params => header_params,
-        :query_params => query_params,
-        :form_params => form_params,
-        :body => post_body,
-        :auth_names => auth_names,
-        :return_type => 'UserCompletionRateResponse')
-      if @api_client.config.debugging
-        @api_client.config.logger.debug "API called: AccountsApi#get_completion_rate_for_template_owner\nData: #{data.inspect}\nStatus code: #{status_code}\nHeaders: #{headers}"
       end
       return data, status_code, headers
     end
@@ -2722,6 +2634,8 @@ module DocuSign_eSign
       query_params[:'include_closed_users'] = options.include_closed_users if !options.include_closed_users.nil?
       query_params[:'permissions'] = options.permissions if !options.permissions.nil?
       query_params[:'start_position'] = options.start_position if !options.start_position.nil?
+      query_params[:'task_source'] = options.task_source if !options.task_source.nil?
+      query_params[:'task_type'] = options.task_type if !options.task_type.nil?
       query_params[:'user_name_substring'] = options.user_name_substring if !options.user_name_substring.nil?
 
       # header parameters